Add "Deactivate" and "Toggle" commands
Adds command line options to send commands to the daemon to deactivate
or toggle the window picker.  Deactivate hides the window picker, if
active, without changing window focus.  Toggle either hides the window
picker if active or shows the window picker if inactive.  This is useful
for binding to a key.