Added some missing stuff
Namely, a couple of links I missed, the blurb from the main google code project page, a link to the downloads page on the Google Code project and the youtube demo link.  I tried to convert the Youtube video into a gif so we can embed it straight into the README on Github but my computer was playing silly buggers.